Add ticket-frontend and ticket-backend folders to fsdFinal

+++ ticket-backend/src/main/java/com/ticket/ticket_backend/repository/TicketRepository.java
@@ -2,8 +2,9 @@ package com.ticket.ticket_backend.repository;
 
 import com.ticket.ticket_backend.model.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.List;
 
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
+public interface TicketRepository extends JpaRepository<Ticket, Integer> {
     List<Ticket> findByGmail(String gmail);
 }
